requirements.txt
python-telegram-bot==20.1
google-api-python-client==2.101.0
google-auth==2.25.0
google-auth-oauthlib==0.5.3
google-auth-httplib2==0.1.0
google-api-core==2.12.0
google-cloud==0.34.0
requests==2.28.1
